refactor(product): Change product dto's to product folder
