+++ services/inventory/src/main/java/com/petek/inventoryService/dto/product/ProductResponse.java
@@ -1,8 +1,10 @@
-package com.petek.inventoryService.dto;
+package com.petek.inventoryService.dto.product;
 
 import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.List;
+
+import com.petek.inventoryService.dto.SupplierResponse;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
